add field pickers, add next button

+++ src/load_verses.rs
@@ -7,15 +7,15 @@ use std::{collections::HashMap, include_str};
 pub struct Verse {
     pub book_id: String,
     pub book_name: String,
-    pub chapter: u32,
-    pub verse: u32,
+    pub chapter: usize,
+    pub verse: usize,
     pub text: String,
 }
 
 #[derive(Debug, Default, Clone)]
 pub struct Book {
     pub book_name: String,
-    pub chapters: HashMap<u32, Vec<Verse>>,
+    pub chapters: HashMap<usize, Vec<Verse>>,
 }
 
 const VERSES_TEXT_RST: &str = include_str!("../bible_translations/rst.json");
@@ -33,7 +33,7 @@ pub fn load_rst_verses() -> Vec<Verse> {
 
 pub fn get_book(book_name: String, verses: Vec<Verse>) -> Book {
     let all_book_verses: Vec<Verse> = verses.into_iter().filter(|v| v.book_name == book_name).collect();
-    let mut chapters: HashMap<u32, Vec<Verse>> = HashMap::new();
+    let mut chapters: HashMap<usize, Vec<Verse>> = HashMap::new();
 
     for verse in all_book_verses {
         chapters.entry(verse.chapter).or_insert(Vec::new()).push(verse);
